Moving to bot directory
