+++ Makefile
@@ -1,10 +1,10 @@
 PHONY: lint reformat
 
 lint:
-	black --check --line-length 100 ./CCC_Bot.py ./cogs/ ./utils/
-	pylint CCC_Bot.py ./utils/ ./cogs/
-	flake8 CCC_Bot.py ./utils/ ./cogs/ --statistics --show-source --max-line-length 100
-	bandit -r ./CCC_Bot.py ./cogs/ ./utils/
+	black --check --line-length 100 ./bot
+	pylint ./bot
+	flake8 ./bot --statistics --show-source --max-line-length 100
+	bandit -r ./bot
 
 reformat:
-	black --line-length 100 ./CCC_Bot.py ./cogs/ ./utils/
+	black --line-length 100 ./bot
