+++ requirements.txt
@@ -1,5 +1,5 @@
-asyncio == 3.4.3
-aiohttp ==3.7.4.post0
-discord.py ==1.7.3
-pandas ==1.2.4
-psycopg2-binary == 2.8.6
+asyncio== 3.4.3
+aiohttp==3.7.4.post0
+discord.py==1.7.3
+pandas==1.2.4
+psycopg2-binary== 2.8.6
